updating the rep. 11th jan..

+++ TimesheetManagement/src/iiitb/timesheet/action/UpdateEmployee.java
@@ -16,7 +16,7 @@ public class UpdateEmployee extends ActionSupport  {
 	private String firstname="";
 	private String lastname="";
 	private String email="";
-	private int phone_num=0;
+	private long phone_num=0;
 	private String join_date;
 	private String role_name;
 	private int emp_id;
@@ -67,11 +67,11 @@ public class UpdateEmployee extends ActionSupport  {
 		this.email = email;
 	}
 
-	public int getPhone_num() {
+	public long getPhone_num() {
 		return phone_num;
 	}
 
-	public void setPhone_num(int phone_num) {
+	public void setPhone_num(long phone_num) {
 		this.phone_num = phone_num;
 	}
